+++ packages/sources/src/components/card/card.content.ts
@@ -1,0 +1,14 @@
+export const cardContent = {
+  label: "Omgevingsplan Nieuwegein",
+  selectable: false,
+  image: undefined,
+  interactions: [
+    {
+      variant: "tertiary",
+      label: "Toon informatie",
+      icon: {
+        icon: "info",
+      },
+    },
+  ],
+};