+++ packages/sources/src/components/card/card.stories-of.ts
@@ -1,7 +1,6 @@
-import { componentArgs } from "../../storybook";
 import { StoriesOfArguments, storiesOfFactory } from "../../storybook/stories-of-factory";
-
 import { CardArgs, cardArgsMapper, cardArgTypes } from "./card.args";
+import { cardContent } from "./card.content";
 import { Card } from "./card.models";
 
 export interface CardTemplates<TemplateFnReturnType> {
@@ -20,38 +19,28 @@ export function storiesOfCard<Implementation, Templates, TemplateFnReturnType>(
   return storiesOfFactory("Card", storiesOfArguments, (stories, templateMapper) => {
     stories.addParameters({
       argTypes: cardArgTypes,
-      args: componentArgs<CardArgs>({
-        label: "Omgevingsplan Nieuwegein",
-        selectable: false,
-        image: undefined,
-        interactions: [
-          {
-            variant: "tertiary",
-            label: "Toon informatie",
-            icon: {
-              icon: "info",
-            },
-          },
-        ],
-      }),
     });
 
     const template = templateMapper<CardArgs>((args, { cardTemplate, content }) =>
       cardTemplate(cardArgsMapper(args, content))
     );
 
-    stories.add("default", template);
+    stories.add("default", template, {
+      args: cardContent,
+    });
 
     stories.add("selectable", template, {
-      args: componentArgs<Pick<CardArgs, "selectable">>({
+      args: {
+        ...cardContent,
         selectable: true,
-      }),
+      },
     });
 
     stories.add("with image", template, {
-      args: componentArgs<Pick<CardArgs, "image">>({
+      args: {
+        ...cardContent,
         image: "images/rectangle1.png",
-      }),
+      },
     });
   });
 }