+++ packages/storybook/src/components/card/card.content.ts
@@ -1,5 +1,3 @@
 import { html } from "lit-html";
 
-export const content = html`<div class="dso-rich-content">
-  Gemeente Nieuwegein lorem ipsum dolor sit amet, consectetur adipiscing elit.
-</div>`;
+export const content = html`<p>Gemeente Nieuwegein lorem ipsum dolor sit amet, consectetur adipiscing elit.</p>`;